requirements.txt
Django==1.5.2
croniter==0.3.3
dj-database-url==0.2.2
dj-static==0.0.5
django-filter==0.7
django-toolbelt==0.0.1
djangorestframework==2.3.7
gunicorn==17.5
psycopg2==2.5.1
python-dateutil==2.1
six==1.3.0
static==0.4
wsgiref==0.1.2
zope.interface==4.0.5
